pyproject.toml
[build-system]
requires = ["sphinx-theme-builder >= 0.2.0a7"]
build-backend = "sphinx_theme_builder"

[tool.sphinx-theme-builder]
node-version = "16.13.2"
theme-name = "pydata_sphinx_theme"
additional-compiled-static-assets = [
  "webpack-macros.html",
  "vendor/",
  "styles/bootstrap.css",
  "scripts/bootstrap.js",
  "locale/"
]

[project]
name = "pydata-sphinx-theme"
description = "Bootstrap-based Sphinx theme from the PyData community"
dynamic = ["version"]
readme = "README.md"

requires-python = ">=3.7"
dependencies = [
  "sphinx>=4.2",
  "beautifulsoup4",
  "docutils!=0.17.0",
  "packaging",
  "Babel",
  "pygments>=2.7",
  "accessible-pygments"
]

license = { file = "LICENSE" }
maintainers = [
  { name = "Joris Van den Bossche", email = "jorisvandenbossche@gmail.com" },
]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Framework :: Sphinx",
  "Framework :: Sphinx :: Theme",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
doc = [
  "numpydoc",
  "myst-nb",
  "linkify-it-py",  # for link shortening
  "rich",
  "sphinxext-rediraffe",
  "sphinx-sitemap",
  # For examples section
  "ablog>=0.11.0rc2",
  "jupyter_sphinx",
  "pandas",
  "plotly",
  "matplotlib",
  "numpy",
  "xarray",
  "sphinx-copybutton",
  "sphinx-design",
  "sphinx-togglebutton",
  "sphinxcontrib-youtube",
  "sphinx-favicon>=1.0.1",
  # Install nbsphinx in case we want to test it locally even though we can't load
  # it at the same time as MyST-NB.
  "nbsphinx",
  "ipyleaflet",
  "colorama",
]
test = [
  "pytest",
  "pytest-cov",
  "pytest-regressions",
  "pytest-playwright",
  "codecov",
]
dev = [
  "pyyaml",
  "pre-commit",
  "nox",
  "pydata-sphinx-theme[doc,test]",
]

[project.entry-points]
"sphinx.html_themes" = { pydata_sphinx_theme = "pydata_sphinx_theme" }

[tool.doc8]
ignore = ["D001"] # we follow a 1 line = 1 paragraph style
